+++ yudao-module-system/yudao-module-system-biz/src/main/java/cn/iocoder/yudao/module/system/service/tenantconfig/TenantConfigService.java
@@ -1,10 +1,15 @@
 package cn.iocoder.yudao.module.system.service.tenantconfig;
 
-import java.util.*;
-import javax.validation.*;
-import cn.iocoder.yudao.module.system.controller.admin.tenantconfig.vo.*;
-import cn.iocoder.yudao.module.system.dal.dataobject.tenantconfig.TenantConfigDO;
 import cn.iocoder.yudao.framework.common.pojo.PageResult;
+import cn.iocoder.yudao.module.system.controller.admin.tenantconfig.vo.TenantConfigCreateReqVO;
+import cn.iocoder.yudao.module.system.controller.admin.tenantconfig.vo.TenantConfigExportReqVO;
+import cn.iocoder.yudao.module.system.controller.admin.tenantconfig.vo.TenantConfigPageReqVO;
+import cn.iocoder.yudao.module.system.controller.admin.tenantconfig.vo.TenantConfigUpdateReqVO;
+import cn.iocoder.yudao.module.system.dal.dataobject.tenantconfig.TenantConfigDO;
+
+import javax.validation.Valid;
+import java.util.Collection;
+import java.util.List;
 
 /**
  * 租户参数配置 Service 接口
@@ -19,7 +24,7 @@ public interface TenantConfigService {
      * @param createReqVO 创建信息
      * @return 编号
      */
-    Integer createTenantConfig(@Valid TenantConfigCreateReqVO createReqVO);
+    Long createTenantConfig(@Valid TenantConfigCreateReqVO createReqVO);
 
     /**
      * 更新租户参数配置
@@ -33,7 +38,7 @@ public interface TenantConfigService {
      *
      * @param id 编号
      */
-    void deleteTenantConfig(Integer id);
+    void deleteTenantConfig(Long id);
 
     /**
      * 获得租户参数配置
@@ -41,7 +46,7 @@ public interface TenantConfigService {
      * @param id 编号
      * @return 租户参数配置
      */
-    TenantConfigDO getTenantConfig(Integer id);
+    TenantConfigDO getTenantConfig(Long id);
 
     /**
      * 获得租户参数配置列表
@@ -49,7 +54,7 @@ public interface TenantConfigService {
      * @param ids 编号
      * @return 租户参数配置列表
      */
-    List<TenantConfigDO> getTenantConfigList(Collection<Integer> ids);
+    List<TenantConfigDO> getTenantConfigList(Collection<Long> ids);
 
     /**
      * 获得租户参数配置分页
